\name{LS2Wstat-package}
\alias{LS2Wstat-package}
\alias{LS2Wstat}
\docType{package}
\title{
Stationarity testing for locally stationary wavelet fields
}
\description{
This package contains functions for testing for stationarity within images, specifically locally stationary wavelet (LS2W) fields. In addition the package contains functions for implementing quadtree image decompositions, as well as code for simulating LS2W processes for a given spectral structure.
}
\details{
\tabular{ll}{
Package: \tab LS2Wstat\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2018-05-21\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Sarah L. Taylor and Matt Nunes

Maintainer: Matthew Nunes <nunesrpackages@gmail.com>

}
\references{
For details on testing LS2W fields for stationarity, see

Taylor, S.L., Eckley, I.A., and Nunes, M.A. (2014) A Test of Stationarity for Textured Images. \emph{Technometrics}, \bold{56} (3), 291-301.\cr

For further information on LS2W processes, see

Eckley, I.A., Nason, G.P., and Treloar, R.L. (2010) Locally stationary wavelet fields with application to the modelling and analysis of image texture \emph{Journal of the Royal Statistical Society Series C}, \bold{59}, 595-616.
}
\keyword{ package }

